pyproject.toml
[project]
name = "temp-group-moder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "langdetect>=1.0.9",
    "python-dotenv>=1.1.1",
]
